Testcase: merged activity and event
